+++ src/spellbot/data/guild_data.py
@@ -24,5 +24,6 @@ class GuildData:
     notice: str | None
     suggest_voice_category: str | None
     enable_mythic_track: bool
+    active: bool = True
     channels: list[ChannelData] = field(default_factory=list)
     awards: list[GuildAwardData] = field(default_factory=list)