+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds an `active` column to guilds to filter them out of tasks.
+
 ## [v19.1.5](https://github.com/lexicalunit/spellbot/releases/tag/v19.1.5) - 2026-05-22
 
 ### Added